requirements.txt
asgiref==3.3.4
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
flake8==3.9.2
gunicorn==20.1.0
mccabe==0.6.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
sqlparse==0.4.1
